discussion: empêcher la suppression de fils avec commentaires
C'est un peu problématique d'être en mesure d'effacer un fil entier avec
tous les commentaires seulement parce qu'on a commencé le fil.

S'il y a un commentaire dans le fil, on ne peut plus le supprimer.

+++ src/tn2app/views/__init__.py
@@ -222,6 +222,8 @@ class DiscussionEdit(LoginRequiredMixin, UpdateView):
     def post(self, request, *args, **kwargs):
         if 'delete' in request.POST:
             discussion = self.get_object()
+            if not discussion.can_delete():
+                raise PermissionDenied()
             group = discussion.group
             discussion.delete()
             return HttpResponseRedirect(group.get_absolute_url())
